pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
